Tutorial part 5: Monsters on the map take a turn when player does and respond to attacks
They don't move yet, but they write a comment to STDOUT when the player moves and when the player "attacks". Also, fixed the redraw issue to monsters outside of the FOV don't materialize when the character moves. I was still using put_char_ex for GameElement#clear which was using explicit colors, and didn't respect the map's colors